refactor: Extract conversion to interval seconds from various types
The procedure to convert a given time object to a float containing
the "seconds until the next run" was extracted into an auxiliary
top-level function named `_to_interval_seconds`.

Also, `_put`'s "prep phase" was refactored as follows. A new local
identifier named `interval` was created to make its meaning clearer
(instead of always reusing `next_t`, which was confusing). The
way in which the interval is fetched (and checked for `None`)
has also been simplified. Finally, `next_t`'s last assignment is
the actual timestamp in which the job is to be run, which is
either `previous_t` (if available) or the current time plus
the amount stored in `interval`.

Finally, `last_t` was renamed to `previous_t` to avoid confusion
with upcoming `last` kwarg.