feat: Add `last` kwarg
Added a `last` kwarg to `run_repeating` that specifies when
the job should last be run. The checking is done before executing
the job, but with a small delay buffer, so that if `last`
is an exact multiple of `interval`, the last job is still
run, even though the actual time is slightly after the
theoretical finish time.