Fixed obsolete documentation.
Removed warnings of the thumbnail_url parameters, regarding the deprecated parameter thumb, as well as the optional annotation in the docs.
Instead, added 'versionchanged' everywhere pointing to 20.5, the version where the  thumb parameter was removed.
Added a new substitutions ('thumbnail_url_mandatory') to global.rst, used in all these 'versionchanged'.
Also removed the now obsolete 'Raises' paragraphs from all classes.
For file inlinequeryresultvideo.py it was also necessary to remove the optional annotation as well as the warning from title parameter. this parameter had to be made optional earlier, to be able to made thumb_url/thumbnail_url optional.

Also had a look at the other inlinequeryresult* files, but there seems everything to be all right.